requirements.txt
fastapi
uvicorn[standard]
scikit-learn
pandas
joblib
pillow
numpy
python-multipart
